Corretti tipi per evitare eventuali overflow

+++ esp32/src/encoder/Encoder.hpp
@@ -6,10 +6,10 @@
 class Encoder {
     public:
         virtual ~Encoder() {}
-        virtual int8_t getPosition() = 0;
-        virtual uint16_t getClockwiseRevolutions() = 0;
-        virtual uint16_t getCounterclockwiseRevolutions() = 0;
-        virtual uint16_t getRevolutions() = 0;
+        virtual int16_t getPosition() = 0;
+        virtual uint32_t getClockwiseRevolutions() = 0;
+        virtual uint32_t getCounterclockwiseRevolutions() = 0;
+        virtual uint32_t getRevolutions() = 0;
         virtual void begin() = 0;
 };
